requirements.txt
streamlit
openai>=1.0.0
python-dotenv
sqlparse
altair>=4.2.2